// packages/sdk/src/constants/local.ts
// Auto-generated by scripts/deploy.ts at 2026-05-30T20:17:13.601Z
import type { ContractAddresses } from "../types.js";

export const LOCAL_ADDRESSES: ContractAddresses = {
  "registry": "5EXroPWCsgdDtg2v4drzetFRioxaFLvmF4XCVfyS9uQB7R8W",
  "resolver": "5Dg6Ck54CFbGtbZjdm5L7qZqzei5UBFoToXeu6txXeZbfa3b",
  "reverseRegistrar": "5GJbAXyhwz2akaKe5Wsxt2wC2wr9xnKEcFiZBg1KWtzSw8GM",
  "registrar": "5EG7qdsXbWiaHLuZdZfrsNtpfTjcKt8QxrKQssZtdiszSdPx",
  "attestation": "5GHSxCsP4aJXpaiTdusYAJ2HGCHT2sBqyEaBVoMEhUxB6oe8"
};